+++ lib/acf/theme-options.php
@@ -294,6 +294,28 @@ function tk_add_acf_theme_options()
 					),
 				),
 			),
+            array (
+                'key' => 'field_tk_tab_404',
+                'label' => '404 page',
+                'type' => 'tab',
+            ),          
+            array (
+                'key' => 'field_tk_404_page_title',
+                'label' => '404 Page (page not found) title',
+                'name' => 'tk_404_page_title',
+                'type' => 'text',
+                'placeholder' => 'Page not found',
+            ),
+            array (
+                'key' => 'field_tk_404_page_content',
+                'label' => '404 Page (page not found) content',
+                'name' => 'tk_404_page_content',
+                'type' => 'wysiwyg',
+                'tabs' => 'all',
+                'toolbar' => 'basic',
+                'placeholder' => 'Sorry, the page you are looking for could not be found on this site',
+                'media_upload' => 1,
+            ),
 	    )),
 	    'location' => array (
 	        array (