+++ widgets/widget-featured-content.php
@@ -6,6 +6,8 @@ if(get_sub_field('featured_content_widget_background') == 'grey'){
 	$featured_content_background = "";
 }
 
+$link_url = (get_sub_field('featured_content_widget_link_option') == 'internal') ? get_sub_field('featured_content_widget_internal_link'): get_sub_field('featured_content_widget_external_link');
+
 ?>
 
 <!-- Featured Content Widget-->
@@ -24,27 +26,23 @@ if(get_sub_field('featured_content_widget_background') == 'grey'){
 			            <?php the_sub_field('featured_content_widget_content'); ?>
 			        </div>    
 
-			        <?php if(get_sub_field('featured_content_widget_link_option') == 'internal') { ?>			
-	            		<a class="more" href="<?php the_sub_field('featured_content_widget_internal_link');?>">More</a>
-	            	<?php } elseif(get_sub_field('featured_content_widget_link_option') == 'external'){ ?>		
-	            		<a class="more" href="<?php the_sub_field('featured_content_widget_external_link');?>">More</a>
+			        <?php if ( $link_url ) { ?>			
+	            		<a class="more" href="<?php echo $link_url; ?>">More</a>
 	            	<?php } ?>
 	            	
 			    </div>
 			    <div class="col-sm-4 col-md-6">			    	
-			    	<?php if(get_sub_field('featured_content_widget_image')){ ?>
+			    	<?php if(get_sub_field('featured_content_widget_media_type') != "video" && get_sub_field('featured_content_widget_image')): ?>
 			        <div class="rs-img rs-img-2-1" style="background-image: url('<?php the_sub_field('featured_content_widget_image'); ?>');">
-				        <?php if(get_sub_field('featured_content_widget_link_option') == 'internal') { ?>			
-		            		<a href="<?php the_sub_field('featured_content_widget_internal_link');?>">
-		            	<?php } elseif(get_sub_field('featured_content_widget_link_option') == 'external'){ ?>		
-		            		<a href="<?php the_sub_field('featured_content_widget_external_link');?>">
-		            	<?php } ?>
-			            		<img src="<?php the_sub_field('featured_content_widget_image'); ?>" alt="<?php the_sub_field('featured_content_widget_heading'); ?>">
-			            <?php if(get_sub_field('featured_content_widget_link_option') != 'no-link') { ?>
-			            	</a>
-			            <?php } ?>			
+				        <?php if ( $link_url ) : ?><a href="<?php echo $link_url; ?>"><?php endif; ?>
+			            	<img src="<?php the_sub_field('featured_content_widget_image'); ?>" alt="<?php the_sub_field('featured_content_widget_heading'); ?>">
+			            <?php if ( $link_url ) : ?></a><?php endif; ?>			
 			        </div>
-			        <?php } ?>
+			        <?php elseif(get_sub_field('featured_content_widget_media_type') == "video" && get_sub_field('featured_content_widget_video')): ?>
+                    <div class="embed-responsive embed-responsive-16by9">
+                        <?php echo get_sub_field('featured_content_widget_video'); ?>
+                    </div>
+                    <?php endif; ?>
 			    </div>
 			</div>